requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
biopython==1.78
matplotlib==3.4.2
seaborn==0.11.1
jupyter==1.0.0
snakemake==6.10.0
pytest==6.2.4
statsmodels>=0.12.0
requests>=2.25.0
tqdm>=4.60.0